+++ haskell/ngx-export-distribution/NgxExport/Distribution.hs
@@ -38,8 +38,10 @@ import Distribution.Types.BuildInfo
 import Distribution.Types.Library
 import Distribution.Verbosity
 import Distribution.Pretty
+import System.Directory
 import System.FilePath
 import Control.Arrow
+import Control.Monad
 import Data.Maybe
 
 -- $usage
@@ -193,6 +195,16 @@ buildSharedLib verbosity desc lbi flags = do
                        nameSo = addExtension name "so"
                    in (nameSo, [addExtension name "hs", "-o", nameSo])
                   ) (, []) lib
+    (extraSourceFile, extraSourceFileDoesNotExist) <-
+        if null extraGhcOptions
+            then return (Nothing, False)
+            else do
+                let file = head extraGhcOptions
+                (Just file, ) . not <$> doesFileExist file
+    when extraSourceFileDoesNotExist $
+        ioError $ userError $
+            "File " ++ fromJust extraSourceFile ++ " does not exist,\
+            \ you may want to specify input and output files in --ghc-options"
     ghcP <- fst <$> requireProgram verbosity ghcProgram (withPrograms lbi)
     let ghcR = programInvocation ghcP $
             ["-dynamic", "-shared", "-fPIC", "-flink-rts"] ++