+++ haskell/ngx-export-distribution/Changelog.md
@@ -12,7 +12,7 @@
 
 ### 0.1.1.0
 
-- Support only Cabal versions >= 3.0.0.0.
+- Support only Cabal versions >= *3.0.0.0*.
 
 ### 0.1.0.0